Graphical version of Steggers department program
Added multiple filetypes, that can be changed dynamically.
Added model that binds to the FXML view
Added addAll to the file interface
Using 3-layered architecture

…
Still needs work on the exception handling system. This is still
inconsistent and exceptions are not handled correctly, just shown in
the System.out through the logger.