Changed Department BE to FX Properties
Example of a BE using FXProperties instead of simple datatypes.

Warning: This makes the BE dependent on the javafx.beans.property
package, which is application specific to JavaFX applications.

I would recommend the javafx properties are limited to the GUI.model
only. However in practise this would mean that we need to copy a lot of
instances and have a cache version in the model.

Try to limit the usage to GUI bindings only.